revoked token

+++ youtravel/src/main/java/com/youcode/youtravel/repositories/TokenRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface TokenRepository extends JpaRepository<Token, Integer> {
+public interface TokenRepository extends JpaRepository<Token, Long> {
 
    // select t from Token t inner join User u on t.user.uid = u.uid
    // where t.user.uid = :userId and t.loggedOut = false
@@ -18,8 +18,6 @@ public interface TokenRepository extends JpaRepository<Token, Integer> {
 
 
     @Query("""
-        
-        
         select t from Token t inner join User c on t.user.uid = c.uid
               where c.uid = :userId and (t.expired = false OR t.revoked = false)
     """)
